car_admin/requirements.txt
django-ckeditor==5.6.1
pillow==5.3.0
